src/transform-hub/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
lxml==5.2.1
requests==2.31.0
python-multipart==0.0.9
cachetools==5.3.3
dnspython==2.6.1
